requirements.txt
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
dialogflow==1.1.0
Flask==1.1.2
google-api-core==1.24.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
grpcio==1.34.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-telegram-bot==12.8
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
tornado==6.1
urllib3==1.26.2
Werkzeug==1.0.1
gunicorn==20.0.4